Code changes for unit-testing (#223)
* Fix compiler warnings when the TCP Window is not used (#124)

* Fix warnings when TCP window is not used

* Uncrustify

* Move local variables to inner loop in prvNetworkInterfaceInput() (#144)


* Update litani submodule (#147)


* Fix doxygen check (#149)

* Update doxygen version

* update the config file

* TCP_WIN: fix compile warning on x86_64 (#148)

* TCP_WIN: fix compile warning on x86_64

Fix the following warning when building for 64 bit:

warning: conversion from ‘long unsigned int’ to ‘uint32_t’ {aka ‘unsigned int’} changes value from ‘18446744073709551615’ to ‘4294967295’ [-Woverflow]
             uint32_t ulReturn = ~0UL;
                                 ^

* Update FreeRTOS_TCP_WIN.c



* fix deprecated volatile compound assignment (#152)

* fixed deprecated volatile compound assignment

C++20 deprecates some undefined or unclear use cases of 'volatile' like
compound assignments and compliant compilers warn about those deprecated
operations.

In vStreamBufferMoveMid the deprecated compound assignment and other
direct accesses to volatile 'StreamBuffer_t->uxMid' is replaced using a
local variable stored back when done.

* Uncrustify


* FreeRTOS_ARP.c : store local addresses only (#120)

* FreeRTOS_ARP.c : store local addresses only

* Added the function xARPWaitResolution()

* Added an entry to lexicon.txt.

* Ran Uncrustify

* Update unit test file

* Update

* Declared xARPWaitResolution() in FreeRTOS_IP.h

* Compare the result of xIsCallingFromIPTask() with pdFALSE in stead of 0


* Remove unnecessary #ifndef (#186)

* Add entropy

* remove warning

* Remove unnecessary ifndef

* Remove unwanted changes

* Don't Fragment Flags patch. (#179)

* Moves all IP flag defines in FreeRTOS_IP_Private.h so that they are accessible to all protocols
Adds definitions for the IP fragmentation flags
Modifies the fragmentation check for incoming frames to drop both the first and later fragments.
Sets the "don't fragment" flag for all outgoing IP frames ( ICMP, DNS, UDP, TCP )
Removes ipGET_UDP_PAYLOAD_OFFSET_FOR_FRAGMENT as it appears obsolete. The stack never outputs fragments.

* Uncrustified

* Uncrustify

* Fixes the fragment offset and fragmentation flags masks ( 0x0FFF and 0xF000 -> 0x1FFF and 0xE000 )
Adds a configuration define ( ipconfigADVERTISE_DONT_FRAGMENT_FLAG ) as suggested by htibosch with a default value of zero for backwards compatibility
Updates the comment that explains the discarding of incoming fragments as discussed with Aniruddha Kanhere

* Adds the 'U' qualifier as requested by hs2gh
fixes a typo in FreeRTOSIPConfigDefaults.h

* Shortens the comment in FreeRTOSIPConfigDefaults as per htibosch's suggestion.

* Renames ipconfigADVERTISE_DONT_FRAGMENT to ipconfigFORCE_IP_DONT_FRAGMENT

* same as last commit, simply forgot to save this before pushing.


* fix IP buffer padding check on 64bit (#146)

* fix IP buffer padding check on 64bit

On 64 bit systems, FreeRTOS_IPInit() would assert
ipconfigBUFFER_PADDING was equal to 14 to "make sure there
is enough space in pucEthernetBuffer to store a pointer."

This prevents the driver from requesting additional
padding, so make the assert greater than or equal to 14.

Also use the final ipBUFFER_PADDING value instead of
ipconfigBUFFER_PADDING, which is probably what was
intended?

* Update after comments


* Update readme.md (#189)

Just fixing the "table of 3 types of STH32H7" so that it renders in github webpage.  I had a tough time reading that table until I looked at the md source.
You could also just put code fences around it:
~~~
/**
 * RAM area	H747	H743	H742	Location
 * ------------------------------------------------
 * DTCM		128k	128k	128k	0x20000000
 * AXI-SRAM	511k	511k	384k	0x24000000
 *
 * SRAM1	128k	128k	32k		0x30000000
 * SRAM2	128k	128k	16k		0x30020000
 * SRAM3	32k		32k	 	-		0x30040000
 * SRAM4	64k		64k		64k		0x38000000
 * Backup   SRAM	4k		4k	4k	0x38800000
 */
~~~


* Update files referencing aws_application_version.h to use iot_application_version.h (#188)

* Update files referencing aws_application_version.h to use iot_application_version.h

* Remove pic32 ethernet _Command_Version function to remove dependency on iot_application_version.h from amazon-freertos repository.

* Remove function defs from header files (#190)

* Add entropy

* remove warning

* Remove function defs from headers

* Some corrections

* More fixes and uncrustify

* Remove the BaseType min function

* Doxygen

* Fix one CBMC proof

* More cbmc proof fixes

* More cbmc fixes

* Some doxygen additions

* Update last CBMC proof

* Doxygen comments

* Doxygen updates

* Doxygen and spell check

* Spell check and unit-test

* Unit test fix

* Update after comments

* Update 2 after comments

* Move function around

* Uncrustify

* Update after comments


* Do not release a network buffer if it equals to NULL (#191)


* Circumvent Qemu MPS2 networking bug (#142)

* Add support for MPS2 networking with lan9118/lan9220

* Fix uncrustify errors

* Enable network interrupt handling

* Add network interrupt support to Qemu MPS2 AN385

* Fix function comment

* Fix Uncrustify errors

* Fix Uncrustify errors

* Fix Uncrustify Errors

* Fix typo

* Cirumvent Qemu MPS2 network bug

* Remove commented code, add doxygen comment

* Add a project for static analysis (#195)

* Add entropy

* remove warning

* Remove unwanted changes

* Update tcp_mem_stats.c

* Add Coverity

* Remove unused files

* Add some features

* clean up

* More clean up

* Unwanted additions removal

* Clean up

* Add 32-bit compile option

* Update after comments

* Uncrustify

* Create uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Update uncrustify.yml

* Make some code changes as a precursor for unit-testing

* remove unused file

* Fixed failing checks

* Fixed CBMC proof
