+++ .github/lexicon.txt
@@ -587,6 +587,7 @@ pcipaddress
 pcmessage
 pcname
 pcnetworkbuffer
+pcrequestedname
 pcs
 pcsource
 pcto
