+++ FreeRTOS_DNS.c
@@ -50,76 +50,6 @@
 /* Exclude the entire file if DNS is not enabled. */
 #if ( ipconfigUSE_DNS != 0 )
 
-    #if ( ipconfigBYTE_ORDER == pdFREERTOS_LITTLE_ENDIAN )
-        #define dnsDNS_PORT             0x3500U  /**< Little endian: Port used for DNS. */
-        #define dnsONE_QUESTION         0x0100U  /**< Little endian representation of a DNS question.*/
-        #define dnsOUTGOING_FLAGS       0x0001U  /**< Little endian representation of standard query. */
-        #define dnsRX_FLAGS_MASK        0x0f80U  /**< Little endian:  The bits of interest in the flags field of incoming DNS messages. */
-        #define dnsEXPECTED_RX_FLAGS    0x0080U  /**< Little Endian: Should be a response, without any errors. */
-    #else
-        #define dnsDNS_PORT             0x0035U  /**< Big endian: Port used for DNS. */
-        #define dnsONE_QUESTION         0x0001U  /**< Big endian representation of a DNS question.*/
-        #define dnsOUTGOING_FLAGS       0x0100U  /**< Big endian representation of standard query. */
-        #define dnsRX_FLAGS_MASK        0x800fU  /**< Big endian: The bits of interest in the flags field of incoming DNS messages. */
-        #define dnsEXPECTED_RX_FLAGS    0x8000U  /**< Big endian: Should be a response, without any errors. */
-
-    #endif /* ipconfigBYTE_ORDER */
-
-/** @brief The maximum number of times a DNS request should be sent out if a response
- * is not received, before giving up. */
-    #ifndef ipconfigDNS_REQUEST_ATTEMPTS
-        #define ipconfigDNS_REQUEST_ATTEMPTS    5
-    #endif
-
-/** @brief If the top two bits in the first character of a name field are set then the
- * name field is an offset to the string, rather than the string itself. */
-    #define dnsNAME_IS_OFFSET    ( ( uint8_t ) 0xc0 )
-
-/* NBNS flags. */
-    #if ( ipconfigUSE_NBNS == 1 )
-        #define dnsNBNS_FLAGS_RESPONSE        0x8000U /**< NBNS response flag. */
-        #define dnsNBNS_FLAGS_OPCODE_MASK     0x7800U /**< NBNS opcode bitmask. */
-        #define dnsNBNS_FLAGS_OPCODE_QUERY    0x0000U /**< NBNS opcode query. */
-    #endif /* ( ipconfigUSE_NBNS == 1 ) */
-
-/* Host types. */
-    #define dnsTYPE_A_HOST    0x01U /**< DNS type A host. */
-    #define dnsCLASS_IN       0x01U /**< DNS class IN (Internet). */
-
-    #ifndef _lint
-        /* LLMNR constants. */
-        #define dnsLLMNR_TTL_VALUE           300000UL /**< LLMNR time to live value. */
-        #define dnsLLMNR_FLAGS_IS_REPONSE    0x8000U  /**< LLMNR flag value for response. */
-    #endif /* _lint */
-
-/* NBNS constants. */
-    #if ( ipconfigUSE_NBNS != 0 )
-        #define dnsNBNS_TTL_VALUE               3600UL  /**< NBNS TTL: 1 hour valid. */
-        #define dnsNBNS_TYPE_NET_BIOS           0x0020U /**< NBNS Type: NetBIOS. */
-        #define dnsNBNS_CLASS_IN                0x01U   /**< NBNS Class: IN (Internet). */
-        #define dnsNBNS_NAME_FLAGS              0x6000U /**< NBNS name flags. */
-        #define dnsNBNS_ENCODED_NAME_LENGTH     32      /**< NBNS encoded name length. */
-
-/** @brief If the queried NBNS name matches with the device's name,
- * the query will be responded to with these flags: */
-        #define dnsNBNS_QUERY_RESPONSE_FLAGS    ( 0x8500U )
-    #endif /* ( ipconfigUSE_NBNS != 0 ) */
-
-/** @brief Flag DNS parsing errors in situations where an IPv4 address is the return
- * type. */
-    #define dnsPARSE_ERROR    0UL
-
-    #ifndef _lint
-        #if ( ipconfigUSE_DNS_CACHE == 0 )
-            #if ( ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY != 1 )
-                #error When DNS caching is disabled, please make ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY equal to 1.
-            #endif
-        #endif
-    #endif
-
-/** @brief Define the ASCII value of '.' (Period/Full-stop). */
-    #define ASCII_BASELINE_DOT    46U
-
 /*
  * Create a socket and bind it to the standard DNS port number.  Return the
  * the created socket - or NULL if the socket could not be created or bound.
@@ -211,19 +141,7 @@
                                               uint32_t ulTTL,
                                               BaseType_t xLookUp );
 
-        typedef struct xDNS_CACHE_TABLE_ROW
-        {
-            uint32_t ulIPAddresses[ ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY ]; /* The IP address(es) of an ARP cache entry. */
-            char pcName[ ipconfigDNS_CACHE_NAME_LENGTH ];                    /* The name of the host */
-            uint32_t ulTTL;                                                  /* Time-to-Live (in seconds) from the DNS server. */
-            uint32_t ulTimeWhenAddedInSeconds;
-            #if ( ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY > 1 )
-                uint8_t ucNumIPAddresses;
-                uint8_t ucCurrentIPAddress;
-            #endif
-        } DNSCacheRow_t;
-
-        static DNSCacheRow_t xDNSCache[ ipconfigDNS_CACHE_ENTRIES ];
+        _static DNSCacheRow_t xDNSCache[ ipconfigDNS_CACHE_ENTRIES ];
 
 /* Utility function: Clear DNS cache by calling this function. */
         void FreeRTOS_dnsclear( void )
@@ -238,21 +156,6 @@
     #endif /* ipconfigUSE_LLMNR == 1 */
 
 /*-----------------------------------------------------------*/
-
-/* Below #include just tells the compiler to pack the structure.
- * It is included in to make the code more readable */
-    #include "pack_struct_start.h"
-    struct xDNSMessage
-    {
-        uint16_t usIdentifier;    /**< Query identifier. Used to match up replies to outstanding queries. */
-        uint16_t usFlags;         /**< Flags. */
-        uint16_t usQuestions;     /**< Number of questions asked in this query. */
-        uint16_t usAnswers;       /**< Number of answers being provided in this query. */
-        uint16_t usAuthorityRRs;  /**< Authoritative name server resource records. */
-        uint16_t usAdditionalRRs; /**< Additional resource records.*/
-    }
-    #include "pack_struct_end.h"
-    typedef struct xDNSMessage DNSMessage_t;
 
 /**
  * @brief Utility function to cast pointer of a type to pointer of type DNSMessage_t.
@@ -297,18 +200,6 @@
         return ( DNSTail_t * ) pvArgument;
     }
 
-/* DNS answer record header. */
-    #include "pack_struct_start.h"
-    struct xDNSAnswerRecord
-    {
-        uint16_t usType;       /**< Type of DNS answer record. */
-        uint16_t usClass;      /**< Class of DNS answer record. */
-        uint32_t ulTTL;        /**< Number of seconds the result can be cached. */
-        uint16_t usDataLength; /**< Length of the data field. */
-    }
-    #include "pack_struct_end.h"
-    typedef struct xDNSAnswerRecord DNSAnswerRecord_t;
-
 /**
  * @brief Utility function to cast pointer of a type to pointer of type DNSAnswerRecord_t.
  *
@@ -319,21 +210,8 @@
         return ( DNSAnswerRecord_t * ) pvArgument;
     }
 
-    #if ( ipconfigUSE_LLMNR == 1 )
 
-        #include "pack_struct_start.h"
-        struct xLLMNRAnswer
-        {
-            uint8_t ucNameCode;    /**< Name type. */
-            uint8_t ucNameOffset;  /**< The name is not repeated in the answer, only the offset is given with "0xc0 <offs>" */
-            uint16_t usType;       /**< Type of the Resource record. */
-            uint16_t usClass;      /**< Class of the Resource record. */
-            uint32_t ulTTL;        /**< Seconds till this entry can be cached. */
-            uint16_t usDataLength; /**< Length of the address in this record. */
-            uint32_t ulIPAddress;  /**< The IP-address. */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xLLMNRAnswer LLMNRAnswer_t;
+    #if ( ipconfigUSE_LLMNR == 1 )
 
 /**
  * @brief Utility function to cast pointer of a type to pointer of type LLMNRAnswer_t.
@@ -349,37 +227,6 @@
     #endif /* ipconfigUSE_LLMNR == 1 */
 
     #if ( ipconfigUSE_NBNS == 1 )
-
-        #include "pack_struct_start.h"
-        struct xNBNSRequest
-        {
-            uint16_t usRequestId;                          /**< NBNS request ID. */
-            uint16_t usFlags;                              /**< Flags of the DNS message. */
-            uint16_t ulRequestCount;                       /**< The number of requests/questions in this query. */
-            uint16_t usAnswerRSS;                          /**< The number of answers in this query. */
-            uint16_t usAuthRSS;                            /**< Number of authoritative resource records. */
-            uint16_t usAdditionalRSS;                      /**< Number of additional resource records. */
-            uint8_t ucNameSpace;                           /**< Length of name. */
-            uint8_t ucName[ dnsNBNS_ENCODED_NAME_LENGTH ]; /**< The domain name. */
-            uint8_t ucNameZero;                            /**< Terminator of the name. */
-            uint16_t usType;                               /**< Type of NBNS record. */
-            uint16_t usClass;                              /**< Class of NBNS request. */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xNBNSRequest NBNSRequest_t;
-
-        #include "pack_struct_start.h"
-        struct xNBNSAnswer
-        {
-            uint16_t usType;       /**< Type of NBNS answer. */
-            uint16_t usClass;      /**< Class of NBNS answer. */
-            uint32_t ulTTL;        /**< Time in seconds for which the answer can be cached. */
-            uint16_t usDataLength; /**< Data length. */
-            uint16_t usNbFlags;    /**< NetBIOS flags 0x6000 : IP-address, big-endian. */
-            uint32_t ulIPAddress;  /**< The IPv4 address. */
-        }
-        #include "pack_struct_end.h"
-        typedef struct xNBNSAnswer NBNSAnswer_t;
 
 /**
  * @brief Utility function to cast pointer of a type to pointer of type NBNSAnswer_t.
@@ -408,18 +255,6 @@
 
     #if ( ipconfigDNS_USE_CALLBACKS == 1 )
 
-/** @brief The structure to hold information for a DNS callback. */
-        typedef struct xDNS_Callback
-        {
-            TickType_t uxRemaningTime;     /**< Timeout in ms */
-            FOnDNSEvent pCallbackFunction; /**< Function to be called when the address has been found or when a timeout has been reached */
-            TimeOut_t uxTimeoutState;      /**< Timeout state. */
-            void * pvSearchID;             /**< Search ID of the callback function. */
-            struct xLIST_ITEM xListItem;   /**< List struct. */
-            char pcName[ 1 ];              /**< 1 character name. */
-        } DNSCallback_t;
-
-
 /**
  * @brief Utility function to cast pointer of a type to pointer of type DNSCallback_t.
  *
@@ -431,7 +266,7 @@
         }
 
 /** @brief The list of all callback structures. */
-        static List_t xCallbackList;
+        _static List_t xCallbackList;
 
 /**
  * @brief Define FreeRTOS_gethostbyname() as a normal blocking call.
@@ -1137,7 +972,7 @@
                         break;
                     }
 
-                    while( ( uxCount-- != 0U ) && ( uxIndex < uxSourceLen ) )
+                    while( uxCount-- != 0U )
                     {
                         if( uxNameLen >= uxDestLen )
                         {
@@ -1157,7 +992,14 @@
                 /* Confirm that a fully formed name was found. */
                 if( uxIndex > 0U )
                 {
-                    if( ( uxNameLen < uxDestLen ) && ( uxIndex < uxSourceLen ) && ( pucByte[ uxIndex ] == 0U ) )
+                    /* Here, there is no need to check for pucByte[ uxindex ] == 0 because:
+                     * When we break out of the above while loop, uxIndex is made 0 thereby
+                     * failing above check. Whenever we exit the loop otherwise, either
+                     * pucByte[ uxIndex ] == 0 (which makes the check here unnecessary) or
+                     * uxIndex >= uxSourceLen (which makes sure that we do not go in the 'if'
+                     * case).
+                     */
+                    if( ( uxNameLen < uxDestLen ) && ( uxIndex < uxSourceLen ) )
                     {
                         pcName[ uxNameLen ] = '\0';
                         uxIndex++;
@@ -1614,7 +1456,11 @@
                 }
 
                 #if ( ipconfigUSE_LLMNR == 1 )
-                    else if( ( usQuestions != ( uint16_t ) 0U ) && ( usType == dnsTYPE_A_HOST ) && ( usClass == dnsCLASS_IN ) && ( pcRequestedName != NULL ) )
+
+                    /* No need to check that pcRequestedName != NULL since is usQuestions != 0, then
+                     * pcRequestedName is assigned with this statement
+                     * "pcRequestedName = ( char * ) pucByte;" */
+                    else if( ( usQuestions != ( uint16_t ) 0U ) && ( usType == dnsTYPE_A_HOST ) && ( usClass == dnsCLASS_IN ) )
                     {
                         /* If this is not a reply to our DNS request, it might an LLMNR
                          * request. */
@@ -2011,11 +1857,14 @@
         {
             BaseType_t x;
             BaseType_t xFound = pdFALSE;
-            uint32_t ulCurrentTimeSeconds = ( xTaskGetTickCount() / portTICK_PERIOD_MS ) / 1000UL;
+            TickType_t xCurrentTickCount = xTaskGetTickCount();
+            uint32_t ulCurrentTimeSeconds;
             uint32_t ulIPAddressIndex = 0;
             static BaseType_t xFreeEntry = 0;
 
             configASSERT( ( pcName != NULL ) );
+
+            ulCurrentTimeSeconds = ( xCurrentTickCount / portTICK_PERIOD_MS ) / 1000UL;
 
             /* For each entry in the DNS cache table. */
             for( x = 0; x < ipconfigDNS_CACHE_ENTRIES; x++ )
@@ -2035,17 +1884,26 @@
                         {
                             #if ( ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY > 1 )
                                 uint8_t ucIndex;
+
                                 /* The ucCurrentIPAddress value increments without bound and will rollover, */
                                 /*  modulo it by the number of IP addresses to keep it in range.     */
                                 /*  Also perform a final modulo by the max number of IP addresses    */
                                 /*  per DNS cache entry to prevent out-of-bounds access in the event */
                                 /*  that ucNumIPAddresses has been corrupted.                        */
+                                if( xDNSCache[ x ].ucNumIPAddresses == 0 )
+                                {
+                                    /* Trying lookup before cache is updated with the number of IP
+                                     * addressed? Maybe an accident. Break out of the loop. */
+                                    break;
+                                }
+
                                 ucIndex = xDNSCache[ x ].ucCurrentIPAddress % xDNSCache[ x ].ucNumIPAddresses;
                                 ucIndex = ucIndex % ( uint8_t ) ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY;
                                 ulIPAddressIndex = ucIndex;
 
                                 xDNSCache[ x ].ucCurrentIPAddress++;
                             #endif /* if ( ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY > 1 ) */
+
                             *pulIP = xDNSCache[ x ].ulIPAddresses[ ulIPAddressIndex ];
                         }
                         else
