+++ test/build-combination/AllDisable/FreeRTOSIPConfig.h
@@ -75,7 +75,7 @@
  * a socket.
  */
 #define ipconfigUSE_DNS_CACHE                      ( 0 )
-#define ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY      ( 6 )
+#define ipconfigDNS_CACHE_ADDRESSES_PER_ENTRY      ( 1 )
 #define ipconfigDNS_REQUEST_ATTEMPTS               ( 2 )
 
 /* The IP stack executes it its own task (although any application task can make
